terminal: start the webserver as soon as possible
Since all the relevant dependancies have been accouted for, we know move
the starting of the main webserver to earlier on in the Run function.
This is in preparation for later when the webserver will be "always-on"
and will be used to server status information about the rest of the LiT
subservers.