multi: let the subServerMgr handle loop, pool & faraday
Now we actually register new sub-servers with the new sub-server
manager. We do this for Loop, Pool & Faraday. With this transition, the
LightningTerminal and rpcProxy structs no longer need to directly manage
these sub-servers. Note that with this transition, any errors
encountered by these sub-servers will no longer cause LiT to crash and
will instead only result in the status of these servers being updated.
Any requests for sub-servers that did not start successfully will result
in an error saying that they are not yet running.