multi: track LiT and LND status
With this commit, we now also track the start-up state of LiT and LiT's
connection to LND. These differ from the other sub-servers (loop, pool &
faraday) because a failure to start LiT or LND is fatal and so should
stop the rest of the start-process, however, we still want the webserver
to continue serving the new Status server so that the UI can query the
start-up status of LND and LiT. So: if any errors occur while
starting/connecting to LND or any other errors occur while starting any
of LiTs other processes, then we throw an error but we dont kill the
main LiT process. The main process is only killed upon receiving a
shutdown signal.