multi: add StopDaemon method to lit
In this commit, a new LitService proto is added and the StopDaemon
method is added to the service. This is added due to the fact that in an
upcoming commit, we want Litd to continue running even if LND cannot be
connected to and so this means that Lit will no longer shutdown if LNDs
StopDaemon method is called.