terminal: register Proxy server with Lit's grpc server
We eventually want a few Lit grpc servers to continue functioning even
if LND did not start up properly. This means that we need to register
these servers with LiT's grpc server instead of using LND's.
Because of this change, we also need to update the itest a bit so that
calls to the proxy server are never expected to succeed if the call is
made via the LND port.