litrpc: add session link field to AddAutopilotSessionRequest
Update the Autopilot service AddAutopilotSessionRequest message to
include a new `group_id` field. Users may set this field to the ID of a
sessino group in order to link the new session to the older ones. The
Session message is also updated with new `group_id` field which will be
populated by the group a session.