session: add a session ID to session key index
In this commit, a new index from session ID to session key (the
serialised local public key of the session) is added to the session
store. This is useful because currently the session store only indexes
by session key but we use session ID to refer to the session throughout
the codebase. This commit adds the new logic which starts writing to the
new index when a session is added via the `CreateSession` method and also
adds a new `GetSessionByID` method which can use the new index to
retrieve the session based on its ID. The next commit will do the
migration to back-fill the index with entries for all existing sessions.