db/sqlc: add session tables and queries
This commit adds all the relational table definitions that we need for
persisting session information. We also add all the queries that we will
need for interacting with these tables.

Since this is the first commit that adds sqlc definitions, we also add
the sqlc.yaml definition file, a sqlc generation script and Makefile
helpers.