make: add ability to specify unit test case
This commit let's us specify commands of the form: `make unit
case=TestNameHere`. Note that this commit by itself is not very useful
in LiT since there are no tests in the main `terminal` package. A later
commit will let us set a package name too and then this will be more
useful.