firewalldb: best effort tight coupling of actions and sessions
In this commit, we do our best to ensure that at least at the time of
action creation, if the session ID is set, then our bbolt actions store
impl will at least first check that the session really does exist. This
also forces us to update our tests in preparation for the SQL store
which will tightly couple the actions and sessions.